requirements.txt
SiderealKundliCraft
pillow